silvia_prototype/requirements.txt
fastapi
uvicorn
scikit-learn
numpy
sentinelhub
Pillow
